Run Code Cleanup on the BodyMonitor library

+++ Source/BBD.BodyMonitor/Buffering/DataBlock.cs
@@ -44,7 +44,7 @@
 
         public long GetHash()
         {
-            return (long)(BlockSize ^ BufferPosition ^ StartIndex ^ EndIndex ^ StartTime.ToUnixTimeMilliseconds() ^ EndTime.ToUnixTimeMilliseconds());
+            return BlockSize ^ BufferPosition ^ StartIndex ^ EndIndex ^ StartTime.ToUnixTimeMilliseconds() ^ EndTime.ToUnixTimeMilliseconds();
         }
     }
 }
